Drag-and-drop rank reorder
This is fairly slow, as it currently does an unnecesary full refresh
of the backlog page. This can much better be done using ajax, but I'd
need to fiddle with the CSS to get the row coloring neatly striped
again, and I'd have to replace the ranks as displayed. Can be done,
but not now.